refactor: move select to seperate file
